requirements.txt
numpy>=1.18.1
pandas>=1.0.3
nltk>=3.4.3
tqdm>=4.45.0
scikit-learn>=0.0
torch>=1.4.0
scipy>=1.3.1
gitpython
transliterate
spacy>=2.1.3,<3
joblib>=0.13.2
python-Levenshtein>=0.12.0
# https://github.com/pytorch/fairseq/issues/3535
fairseq==0.10.2
truecase
sentencepiece>=0.1.83
imohash>=1.0.4
cachetools>=4.0.0
dill>=0.3.0
submitit
faiss-gpu
sacremoses>=0.0.38
nevergrad>=0.4.0.post3
editdistance>=0.5.3
tokenizers>=0.5.2
easse@ git+https://github.com/feralvam/easse.git
kenlm@ git+https://github.com/kpu/kenlm.git
cachier>=1.2.8
huggingface
